Commit #3 : Version deployed to the cloud.
Fully functional. We need to change the styling in future to work with mobile and then add the functionality with preassure by making a new API and adding a new AJAX script to the Preassure page which has this new endpoint -- this will be a repetition of the same process.